# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Edelways project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create-theory": {
            "post": {
                "tags": [
                    "Theories"
                ],
                "summary": "Create new theory",
                "description": "Returns theory data",
                "operationId": "createTheory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTheoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Theory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/theories": {
            "get": {
                "tags": [
                    "Theories"
                ],
                "summary": "Get theories list",
                "description": "Returns theories list",
                "operationId": "getTheoriesData",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "query",
                        "description": "Id of related group",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/theory": {
            "get": {
                "tags": [
                    "Theories"
                ],
                "summary": "Get theory information",
                "description": "Returns theory data",
                "operationId": "getTheoryData",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Theory id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Group"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Theory not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete-theory": {
            "post": {
                "tags": [
                    "Theories"
                ],
                "summary": "Delete theory",
                "description": "Returns success",
                "operationId": "deleteTheory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Theory not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateTheoryRequest": {
                "title": "Create theory request",
                "description": "Create theory request body data",
                "required": [
                    "name",
                    "user_id",
                    "group_id",
                    "document_path"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the new group",
                        "type": "string",
                        "example": "Java"
                    },
                    "group_id": {
                        "title": "Group ID",
                        "description": "Id of group",
                        "type": "integer",
                        "example": "12"
                    },
                    "document_path": {
                        "title": "Document path",
                        "description": "Path of document on the server",
                        "type": "string",
                        "example": "/theory/file.docx"
                    }
                },
                "type": "object"
            },
...
            "DeleteGroupRequest": {
                "title": "Delete request",
                "description": "Delete group request body data",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id of the group",
                        "type": "integer",
                        "example": "3"
                    }
                },
                "type": "object"
            },
...
            "Group": {
                "title": "Group",
                "description": "Group",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new group",
                        "type": "string",
                        "example": "Java"
                    },
                    "code": {
                        "title": "Code",
                        "description": "Random code to connect to group",
                        "type": "string",
                        "example": "GHDFDBU"
                    },
                    "user_id": {
                        "title": "User id",
                        "description": "Id of user who created this group",
                        "type": "string",
                        "example": "23"
                    },
                    "users": {
                        "title": "Users",
                        "description": "Count of joined uses",
                        "type": "integer",
                        "example": "12"
                    },
                    "author": {
                        "title": "Author",
                        "description": "Name of author",
                        "type": "string",
                        "example": "Roman Gurniy"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Group"
                }
            },
...
            "Theory": {
                "title": "Theory",
                "description": "Theory",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the new group",
                        "type": "string",
                        "example": "Java"
                    },
                    "user_id": {
                        "title": "User id",
                        "description": "Id of user who created this group",
                        "type": "string",
                        "example": "23"
                    },
                    "group_id": {
                        "title": "Group ID",
                        "description": "Id of group",
                        "type": "integer",
                        "example": "12"
                    },
                    "document_path": {
                        "title": "Document path",
                        "description": "Path of document on the server",
                        "type": "string",
                        "example": "/theory/file.docx"
                    },
                    "author": {
                        "title": "Author",
                        "description": "Author of theory",
                        "type": "string",
                        "example": "Roman Gurniy"
                    },
                    "date": {
                        "title": "Date",
                        "description": "Date of created the theoury",
                        "type": "string",
                        "example": "March 10, 2022"
                    },
                    "isAdmin": {
                        "title": "IsAdmin",
                        "description": "Returns true is user is admin of this group",
                        "type": "string",
                        "example": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Theory"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}